% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclotomic.R
\docType{methods}
\name{as.cyclotomic}
\alias{as.cyclotomic}
\alias{as.cyclotomic,character-method}
\alias{as.cyclotomic,cyclotomic-method}
\alias{as.cyclotomic,numeric-method}
\alias{as.cyclotomic,bigz-method}
\alias{as.cyclotomic,bigq-method}
\title{Coercion to a 'cyclotomic' object}
\usage{
\S4method{as.cyclotomic}{character}(x)

\S4method{as.cyclotomic}{cyclotomic}(x)

\S4method{as.cyclotomic}{numeric}(x)

\S4method{as.cyclotomic}{bigz}(x)

\S4method{as.cyclotomic}{bigq}(x)
}
\arguments{
\item{x}{a \code{cyclotomic} object or an object yielding a quoted integer or a
quoted fraction after an application of \code{as.character}}
}
\value{
A \code{cyclotomic} object.
}
\description{
Coercion to a 'cyclotomic' object
}
\examples{
as.cyclotomic(2)
as.cyclotomic("1/3")
}
