% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclotomic.R
\name{asComplex}
\alias{asComplex}
\title{Convert cyclotomic number to complex number}
\usage{
asComplex(cyc)
}
\arguments{
\item{cyc}{a cyclotomic number}
}
\value{
A complex number (generally inexact).
}
\description{
Convert a cyclotomic number to a complex number.
}
\examples{
asComplex(zeta(4))
}
