% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{conjugate}
\alias{conjugate}
\title{Conjugate cyclotomic number}
\usage{
conjugate(cyc)
}
\arguments{
\item{cyc}{a cyclotomic number}
}
\value{
A cyclotomic number, the conjugate of \code{cyc}.
}
\description{
Conjugate of a cyclotomic number.
}
\examples{
conjugate(e(4)) # should be -e(4)
}
