% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cyclotomic.R
\name{cyclotomic-unary}
\alias{cyclotomic-unary}
\alias{+,cyclotomic,missing-method}
\alias{-,cyclotomic,missing-method}
\title{Unary operators for cyclotomic objects}
\usage{
\S4method{+}{cyclotomic,missing}(e1, e2)

\S4method{-}{cyclotomic,missing}(e1, e2)
}
\arguments{
\item{e1}{object of class \code{cyclotomic}}

\item{e2}{nothing}
}
\value{
A \code{cyclotomic} object.
}
\description{
Unary operators for cyclotomic objects.
}
