% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt.R
\name{cycSqrt}
\alias{cycSqrt}
\title{Square root as a cyclotomic number}
\usage{
cycSqrt(x)
}
\arguments{
\item{x}{an integer or a \strong{gmp} rational number (\code{bigq} object)}
}
\value{
The square root of \code{x} as a cyclotomic number.
}
\description{
Square root of an integer or a rational number as a cyclotomic
  number.
}
\examples{
cycSqrt(2)
}
