% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkCyclotomic.R
\name{e}
\alias{e}
\title{The primitive n-th root of unity.}
\usage{
e(n)
}
\arguments{
\item{n}{positive integer}
}
\value{
A cyclotomic number.
}
\description{
For example, `e(4) = i` is the primitive 4th root of unity,
  and `e(5) = exp(2*pi*i/5)` is the primitive 5th root of unity.
  In general, `e(n) = exp(2*pi*i/n)`.
}
\examples{
e(4)
}
