% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{imaginaryPart}
\alias{imaginaryPart}
\title{Imaginary part of cyclotomic number.}
\usage{
imaginaryPart(cyc)
}
\arguments{
\item{cyc}{a cyclotomic number}
}
\value{
A cyclotomic number.
}
\description{
The imaginary part of a cyclotomic number.
}
\examples{
imaginaryPart(zeta(9))
}
