% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational.R
\name{isGaussianRational}
\alias{isGaussianRational}
\title{Is the cyclotomic a Gaussian rational?}
\usage{
isGaussianRational(cyc)
}
\arguments{
\item{cyc}{a cyclotomic number}
}
\value{
A Boolean value.
}
\description{
Checks whether a cyclotomic number is a Gaussian rational number.
}
