% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational.R
\name{isRational}
\alias{isRational}
\title{Is the cyclotomic a rational number?}
\usage{
isRational(cyc)
}
\arguments{
\item{cyc}{a cyclotomic number}
}
\value{
A Boolean value.
}
\description{
Checks whether a cyclotomic number is a rational number.
}
\seealso{
\code{\link{maybeRational}}
}
