% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{isReal}
\alias{isReal}
\title{Is the cyclotomic a real number?}
\usage{
isReal(cyc)
}
\arguments{
\item{cyc}{a cyclotomic number}
}
\value{
A Boolean value.
}
\description{
Checks whether a cyclotomic number is a real number.
}
