% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt.R
\name{quadraticRoots}
\alias{quadraticRoots}
\title{Roots of quadratic polynomial}
\usage{
quadraticRoots(a, b, c)
}
\arguments{
\item{a, b, c}{the coefficients of the polynomial}
}
\value{
A list of two cyclotomic numbers, the roots of the polynomial
  \ifelse{html}{\out{ax<sup>2</sup> + bx +c}}{\eqn{ax^2 + bx +c}{ax^2 + bx +c}}.
}
\description{
Roots of a polynomial of degree 2 as cyclotomic numbers.
}
\examples{
library(cyclotomic)
quadraticRoots(a = 1, b = 2, c = -1)
}
