% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugate.R
\name{realPart}
\alias{realPart}
\title{Real part of cyclotomic number.}
\usage{
realPart(cyc)
}
\arguments{
\item{cyc}{a cyclotomic number}
}
\value{
A cyclotomic number.
}
\description{
The real part of a cyclotomic number.
}
\examples{
realPart(zeta(9))
}
