% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry.R
\name{cosDeg}
\alias{cosDeg}
\alias{sinDeg}
\title{Cosine and sine of a rational number}
\usage{
cosDeg(rat)

sinDeg(rat)
}
\arguments{
\item{rat}{an integer number, a \strong{gmp} rational number, or a
fraction given as a character string (e.g. \code{"2/7"})}
}
\value{
A cyclotomic number.
}
\description{
Cosine and sine of a rational number as a cyclotomic number.
}
\details{
\code{cosDeg}, resp. \code{sinDeg}, returns the cosine, resp. the
  sine, of its argument assumed to be given in degrees.
}
\examples{
cosDeg(60)
cosDeg("2/3")^2 + sinDeg("2/3")^2 == 1
}
