% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkCyclotomic.R
\name{zeta}
\alias{zeta}
\title{The primitive n-th root of unity.}
\usage{
zeta(n)
}
\arguments{
\item{n}{a positive integer}
}
\value{
A cyclotomic number.
}
\description{
For example, `zeta(4) = i` is the primitive 4th root of unity,
  and `zeta(5) = exp(2*pi*i/5)` is the primitive 5th root of unity.
  In general, `zeta(n) = exp(2*pi*i/n)`.
}
\examples{
zeta(4)
}
